Fix a few issues
  * Fix typo in cardBackground
  * Fix reboot on uninstall
  * Update strings for hotswap on notification reboot dialog
  * Don't leave OverlaysActivity on Install, Uninstall, or Update
  * Don't update OverlaysActivity while not active Activity

+++ library/src/main/java/com/brit/swiftinstaller/library/installer/rom/OreoRomInfo.kt
@@ -24,6 +24,7 @@ package com.brit.swiftinstaller.library.installer.rom
 import android.content.Context
 import android.content.Intent
 import android.preference.PreferenceManager
+import android.util.Log
 import com.brit.swiftinstaller.library.ui.customize.CustomizeHandler
 import com.brit.swiftinstaller.library.ui.customize.CustomizeSelection
 import com.brit.swiftinstaller.library.ui.customize.PreviewHandler
@@ -99,7 +100,10 @@ open class OreoRomInfo(context: Context) : RomInfo(context) {
     override fun uninstallOverlay(context: Context, packageName: String) {
         if (ShellUtils.isRootAvailable) {
             runCommand("pm uninstall " + getOverlayPackageName(packageName), true)
-            runCommand("pkill $packageName")
+            if (packageName != "com.android.systemui" && packageName != "android") {
+                Log.d("TEST", "killing $packageName")
+                runCommand("pkill $packageName")
+            }
             PreferenceManager.getDefaultSharedPreferences(context)
                     .edit().putBoolean("hotswap", true).apply()
         }
